requirements.txt
websocket_client==1.6.4
